Refactor/added ResumeExperience.ts
